+++ src/module.py
@@ -1,2 +1,2 @@
-def foo(a,b) :
-    return a + b
+def foo(a,b,c) :
+    return a + b + c
