+++ workflow-test-1/src/module.py
@@ -1,0 +1,2 @@
+def foo(a,b) :
+    return a + b
