+++ workflow-test-1/workflow-test-1/.github/workflows/pytest.yaml
@@ -1,0 +1,25 @@
+name : Run Pytest
+on:
+  push:
+    branches :
+      - main
+  pull_request:
+    branches : ["main"]
+    
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.13
+      
+      - name: Install dependencies
+        run: pip install -r requirements.txt
+
+      - name: Run Pytest
+        run: pytest
